cdc,api: restructure API files
