+++ cdc/api/open.go
@@ -46,16 +46,44 @@ const (
 	getOwnerRetryMaxTime = 3
 )
 
-// HTTPHandler is a  HTTPHandler of capture
-type HTTPHandler struct {
+// openAPI provides capture APIs.
+type openAPI struct {
 	capture *capture.Capture
 }
 
-// NewHTTPHandler return a HTTPHandler for OpenAPI
-func NewHTTPHandler(capture *capture.Capture) HTTPHandler {
-	return HTTPHandler{
-		capture: capture,
-	}
+// RegisterOpoenAPIRoutes registers routes for OpenAPI
+func RegisterOpoenAPIRoutes(router *gin.Engine, capture *capture.Capture) {
+	openAPI := openAPI{capture: capture}
+
+	// common API
+	router.GET("/api/v1/status", openAPI.ServerStatus)
+	router.GET("/api/v1/health", openAPI.Health)
+	router.POST("/api/v1/log", SetLogLevel)
+
+	// changefeed API
+	changefeedGroup := router.Group("/api/v1/changefeeds")
+	changefeedGroup.GET("", openAPI.ListChangefeed)
+	changefeedGroup.GET("/:changefeed_id", openAPI.GetChangefeed)
+	changefeedGroup.POST("", openAPI.CreateChangefeed)
+	changefeedGroup.PUT("/:changefeed_id", openAPI.UpdateChangefeed)
+	changefeedGroup.POST("/:changefeed_id/pause", openAPI.PauseChangefeed)
+	changefeedGroup.POST("/:changefeed_id/resume", openAPI.ResumeChangefeed)
+	changefeedGroup.DELETE("/:changefeed_id", openAPI.RemoveChangefeed)
+	changefeedGroup.POST("/:changefeed_id/tables/rebalance_table", openAPI.RebalanceTable)
+	changefeedGroup.POST("/:changefeed_id/tables/move_table", openAPI.MoveTable)
+
+	// owner API
+	ownerGroup := router.Group("/api/v1/owner")
+	ownerGroup.POST("/resign", openAPI.ResignOwner)
+
+	// processor API
+	processorGroup := router.Group("/api/v1/processors")
+	processorGroup.GET("", openAPI.ListProcessor)
+	processorGroup.GET("/:changefeed_id/:capture_id", openAPI.GetProcessor)
+
+	// capture API
+	captureGroup := router.Group("/api/v1/captures")
+	captureGroup.GET("", openAPI.ListCapture)
 }
 
 // ListChangefeed lists all changgefeeds in cdc cluster
@@ -68,7 +96,7 @@ func NewHTTPHandler(capture *capture.Capture) HTTPHandler {
 // @Success 200 {array} model.ChangefeedCommonInfo
 // @Failure 500 {object} model.HTTPError
 // @Router /api/v1/changefeeds [get]
-func (h *HTTPHandler) ListChangefeed(c *gin.Context) {
+func (h *openAPI) ListChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -133,7 +161,7 @@ func (h *HTTPHandler) ListChangefeed(c *gin.Context) {
 // @Success 200 {object} model.ChangefeedDetail
 // @Failure 500,400 {object} model.HTTPError
 // @Router /api/v1/changefeeds/{changefeed_id} [get]
-func (h *HTTPHandler) GetChangefeed(c *gin.Context) {
+func (h *openAPI) GetChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -201,7 +229,7 @@ func (h *HTTPHandler) GetChangefeed(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/changefeeds [post]
-func (h *HTTPHandler) CreateChangefeed(c *gin.Context) {
+func (h *openAPI) CreateChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -246,7 +274,7 @@ func (h *HTTPHandler) CreateChangefeed(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router /api/v1/changefeeds/{changefeed_id}/pause [post]
-func (h *HTTPHandler) PauseChangefeed(c *gin.Context) {
+func (h *openAPI) PauseChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -289,7 +317,7 @@ func (h *HTTPHandler) PauseChangefeed(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/changefeeds/{changefeed_id}/resume [post]
-func (h *HTTPHandler) ResumeChangefeed(c *gin.Context) {
+func (h *openAPI) ResumeChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -337,7 +365,7 @@ func (h *HTTPHandler) ResumeChangefeed(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router /api/v1/changefeeds/{changefeed_id} [put]
-func (h *HTTPHandler) UpdateChangefeed(c *gin.Context) {
+func (h *openAPI) UpdateChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -393,7 +421,7 @@ func (h *HTTPHandler) UpdateChangefeed(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/changefeeds/{changefeed_id} [delete]
-func (h *HTTPHandler) RemoveChangefeed(c *gin.Context) {
+func (h *openAPI) RemoveChangefeed(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -435,7 +463,7 @@ func (h *HTTPHandler) RemoveChangefeed(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router /api/v1/changefeeds/{changefeed_id}/tables/rebalance_table [post]
-func (h *HTTPHandler) RebalanceTable(c *gin.Context) {
+func (h *openAPI) RebalanceTable(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -475,7 +503,7 @@ func (h *HTTPHandler) RebalanceTable(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router /api/v1/changefeeds/{changefeed_id}/tables/move_table [post]
-func (h *HTTPHandler) MoveTable(c *gin.Context) {
+func (h *openAPI) MoveTable(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -526,7 +554,7 @@ func (h *HTTPHandler) MoveTable(c *gin.Context) {
 // @Success 202
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/owner/resign [post]
-func (h *HTTPHandler) ResignOwner(c *gin.Context) {
+func (h *openAPI) ResignOwner(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -549,7 +577,7 @@ func (h *HTTPHandler) ResignOwner(c *gin.Context) {
 // @Success 200 {object} model.ProcessorDetail
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/processors/{changefeed_id}/{capture_id} [get]
-func (h *HTTPHandler) GetProcessor(c *gin.Context) {
+func (h *openAPI) GetProcessor(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -615,7 +643,7 @@ func (h *HTTPHandler) GetProcessor(c *gin.Context) {
 // @Success 200 {array} model.ProcessorCommonInfo
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/processors [get]
-func (h *HTTPHandler) ListProcessor(c *gin.Context) {
+func (h *openAPI) ListProcessor(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -645,7 +673,7 @@ func (h *HTTPHandler) ListProcessor(c *gin.Context) {
 // @Success 200 {array} model.Capture
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/captures [get]
-func (h *HTTPHandler) ListCapture(c *gin.Context) {
+func (h *openAPI) ListCapture(c *gin.Context) {
 	if !h.capture.IsOwner() {
 		h.forwardToOwner(c)
 		return
@@ -680,7 +708,7 @@ func (h *HTTPHandler) ListCapture(c *gin.Context) {
 // @Success 200 {object} model.ServerStatus
 // @Failure 500,400 {object} model.HTTPError
 // @Router	/api/v1/status [get]
-func (h *HTTPHandler) ServerStatus(c *gin.Context) {
+func (h *openAPI) ServerStatus(c *gin.Context) {
 	status := model.ServerStatus{
 		Version: version.ReleaseVersion,
 		GitHash: version.GitHash,
@@ -700,7 +728,7 @@ func (h *HTTPHandler) ServerStatus(c *gin.Context) {
 // @Success 200
 // @Failure 500 {object} model.HTTPError
 // @Router	/api/v1/health [get]
-func (h *HTTPHandler) Health(c *gin.Context) {
+func (h *openAPI) Health(c *gin.Context) {
 	ctx := c.Request.Context()
 
 	if _, err := h.capture.GetOwner(ctx); err != nil {
@@ -741,7 +769,7 @@ func SetLogLevel(c *gin.Context) {
 }
 
 // forwardToOwner forward an request to owner
-func (h *HTTPHandler) forwardToOwner(c *gin.Context) {
+func (h *openAPI) forwardToOwner(c *gin.Context) {
 	ctx := c.Request.Context()
 	// every request can only forward to owner one time
 	if len(c.GetHeader(forWardFromCapture)) != 0 {
