+++ cdc/api/util_test.go
