+++ pkg/cmd/cli/cli_changefeed_list.go
@@ -17,7 +17,7 @@ import (
 	"encoding/json"
 
 	"github.com/pingcap/log"
-	"github.com/pingcap/tiflow/cdc"
+	"github.com/pingcap/tiflow/cdc/api"
 	"github.com/pingcap/tiflow/pkg/cmd/context"
 	"github.com/pingcap/tiflow/pkg/cmd/factory"
 	"github.com/pingcap/tiflow/pkg/cmd/util"
@@ -30,7 +30,7 @@ import (
 // changefeedCommonInfo holds some common used information of a changefeed.
 type changefeedCommonInfo struct {
 	ID      string              `json:"id"`
-	Summary *cdc.ChangefeedResp `json:"summary"`
+	Summary *api.ChangefeedResp `json:"summary"`
 }
 
 // listChangefeedOptions defines flags for the `cli changefeed list` command.
@@ -101,7 +101,7 @@ func (o *listChangefeedOptions) run(cmd *cobra.Command) error {
 			// if no capture is available, the query will fail, just add a warning here
 			log.Warn("query changefeed info failed", zap.String("error", err.Error()))
 		} else {
-			info := &cdc.ChangefeedResp{}
+			info := &api.ChangefeedResp{}
 			err = json.Unmarshal([]byte(resp), info)
 			if err != nil {
 				return err
