+++ pkg/cmd/cli/cli_changefeed_resume.go
@@ -17,7 +17,7 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/pingcap/tiflow/cdc"
+	"github.com/pingcap/tiflow/cdc/api"
 	"github.com/pingcap/tiflow/cdc/model"
 	cmdcontext "github.com/pingcap/tiflow/pkg/cmd/context"
 	"github.com/pingcap/tiflow/pkg/cmd/factory"
@@ -79,7 +79,7 @@ func (o *resumeChangefeedOptions) confirmResumeChangefeedCheck(ctx context.Conte
 		return err
 	}
 
-	info := &cdc.ChangefeedResp{}
+	info := &api.ChangefeedResp{}
 	err = json.Unmarshal([]byte(resp), info)
 	if err != nil {
 		return err
